Extract build-logger.js
