+++ gulpfile.js
@@ -1,3 +1,5 @@
+/* jshint node: true */
+
 var gulp = require('gulp');
 var mocha = require('gulp-mocha');
 var jshint = require('gulp-jshint');
@@ -10,7 +12,7 @@ gulp.task('test', function() {
 });
 
 gulp.task('lint', function() {
-  return gulp.src(['./hookshot.js', './_test/*.js'])
+  return gulp.src(['./*.js', './_test/*.js'])
     .pipe(jshint())
     .pipe(jshint.reporter('default'));
 });
